@@ -3,7 +3,7 @@ import Types
 
 public struct SettingsView: View {
     @Environment(\.presentationMode) var presentationMode
-    @ObservedObject var settingsState: SettingsState
+    @Bindable var settingsState: SettingsState
 
     public init(settingsState: SettingsState) {
         self.settingsState = settingsState
