@@ -2,19 +2,16 @@ import SwiftUI
 import Types
 
 public struct TagListView: View {
-    @ObservedObject var tagState: TagState
-    @Binding var selectionState: ListType?
+    var tagState: TagState
 
     public init(
-        tagState: TagState,
-        selectionState: Binding<ListType?>
+        tagState: TagState
     ) {
         self.tagState = tagState
-        self._selectionState = selectionState
     }
 
     public var body: some View {
-        List(tagState.tags, selection: $selectionState) { tag in
+        List(tagState.tags) { tag in
             NavigationLink(value: ListType.tags(selectedTag: tag)) {
                 TagView(
                     tag: tag,
